Revert changes to store the files location
One slight problem I overlooked was that to be able to open the
config.toml file, you must know the file locations. But the file
location ARE IN THE CONFIG.TOML FILE

So I'll use shared_preferences instead.
I already wasn't too keen on trasnferring save locations between
computers and would prefer if it was reset to the default on new
machines

This reverts commit 27f5fd7e76f959a22da1c0c963f24b4d65bf408c.