Fix tests
It was late yesterday and I'm pretty sure I tested them but apparently
I didn't before committing.
I'll set up an alias to test and commit at the same time so this
doesn't happen again. That's the whole point of tools that prevent you
from breaking things when you're tired.